Lang crate that can be used in backend and frontend. Used it in paragraph example project
