+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "crates/hsluv_macro",
     "crates/route_macro",
+    "crates/lang",
     "crates/static_ref_macro",
     "crates/moon_entry_macros",
     "crates/hsluv",
