+++ crates/zoon/Cargo.toml
@@ -23,6 +23,7 @@ route_macro = { path = "../route_macro", default-features = false, optional = tr
 static_ref_macro = { path = "../static_ref_macro", default-features = false, optional = true }
 moonlight = { path = "../moonlight", features = ["frontend"], optional = true }
 hsluv = { path = "../hsluv", features = ["hsluv_macro"], default-features = false }
+lang = { path = "../lang", default-features = false }
 
 console_error_panic_hook = { version = "0.1.6", default-features = false, optional = true }
 wee_alloc = { version = "0.4.5", default-features = false, optional = true }
