+++ examples/paragraph/backend/Cargo.toml
@@ -11,3 +11,4 @@ edition = "2021"
 
 [dependencies]
 moon = { path = "../../../crates/moon" }
+lang = { path = "../../../crates/lang"}
