+++ examples/paragraph/frontend/Cargo.toml
@@ -17,6 +17,7 @@ wasm-bindgen-test = "0.3.19"
 
 [dependencies]
 zoon = { path = "../../../crates/zoon" }
+lang = { path = "../../../crates/lang"}
 
 [package.metadata.wasm-pack.profile.release]
 wasm-opt = ['-Os']
