+++ examples/paragraph/frontend/src/lib.rs
@@ -1,4 +1,4 @@
-use zoon::Lang;
+use lang::Lang;
 use zoon::{named_color::*, *};
 
 fn root() -> impl Element {
